refactor: extract auth.js
